% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/penguen_grafik.R
\name{penguen_grafik}
\alias{penguen_grafik}
\title{Penguen Turlerine Gore Gaga Uzunlugu Dagilimi}
\usage{
penguen_grafik()
}
\value{
Bir \code{ggplot} nesnesi.
}
\description{
palmerpenguins paketindeki penguins verisini kullanarak tur bazli
gaga uzunlugu (bill_length_mm) dagilimini ggplot2 ile gorsellestirir.
}
\examples{
\dontrun{
  penguen_grafik()
}
}
